fix hapi extensions scope and add missing error tags
