+++ packages/datadog-plugin-hapi/src/server.js
@@ -1,0 +1,184 @@
+'use strict'
+
+const web = require('../../dd-trace/src/plugins/util/web')
+
+function createWrapDispatch (tracer, config) {
+  config = web.normalizeConfig(config)
+
+  return function wrapDispatch (dispatch) {
+    return function dispatchWithTrace (options) {
+      const handler = dispatch.apply(this, arguments)
+
+      return function (req, res) {
+        return web.instrument(tracer, config, req, res, 'hapi.request', () => {
+          return handler.apply(this, arguments)
+        })
+      }
+    }
+  }
+}
+
+function createWrapServer (tracer) {
+  return function wrapServer (server) {
+    return function serverWithTrace (options) {
+      const app = server.apply(this, arguments)
+
+      app.ext = createWrapExt(tracer)(app.ext)
+      app.start = createWrapStart(tracer)(app.start)
+
+      return app
+    }
+  }
+}
+
+function createWrapStart () {
+  return function wrapStart (start) {
+    return function startWithTrace () {
+      this.ext('onPreResponse', onPreResponse)
+
+      return start.apply(this, arguments)
+    }
+  }
+}
+
+function createWrapExt () {
+  return function wrapExt (ext) {
+    return function extWithTrace (events, method, options) {
+      if (typeof events === 'object') {
+        events = wrapEvents(events)
+      } else {
+        method = wrapExtension(method)
+      }
+
+      return ext.call(this, events, method, options)
+    }
+  }
+}
+
+function wrapExtension (method) {
+  return [].concat(method).map(wrapHandler)
+}
+
+function wrapEvents (events) {
+  return [].concat(events).map(event => {
+    return Object.assign({}, event, {
+      method: wrapExtension(event.method)
+    })
+  })
+}
+
+function wrapHandler (handler) {
+  return function (request, h) {
+    if (!request.raw) return handler.apply(this, arguments)
+
+    return web.reactivate(request.raw.req, () => handler.apply(this, arguments))
+  }
+}
+
+function onPreResponse (request, h) {
+  const req = request.raw.req
+
+  web.addError(req, request.response)
+  web.enterRoute(req, request.route.path)
+
+  return reply(request, h)
+}
+
+function reply (request, h) {
+  if (h.continue) {
+    return typeof h.continue === 'function'
+      ? h.continue()
+      : h.continue
+  } else {
+    return h()
+  }
+}
+
+module.exports = [
+  {
+    name: '@hapi/hapi',
+    versions: ['>=17.9'],
+    patch (hapi, tracer, config) {
+      this.wrap(hapi, ['server', 'Server'], createWrapServer(tracer, config))
+    },
+    unpatch (hapi) {
+      this.unwrap(hapi, ['server', 'Server'])
+    }
+  },
+  {
+    name: 'hapi',
+    versions: ['>=17'],
+    patch (hapi, tracer, config) {
+      this.wrap(hapi, ['server', 'Server'], createWrapServer(tracer, config))
+    },
+    unpatch (hapi) {
+      this.unwrap(hapi, ['server', 'Server'])
+    }
+  },
+  {
+    name: 'hapi',
+    versions: ['2 - 7.1', '8 - 16'],
+    patch (hapi, tracer, config) {
+      this.wrap(hapi.Server.prototype, 'start', createWrapStart(tracer, config))
+      this.wrap(hapi.Server.prototype, 'ext', createWrapExt(tracer, config))
+    },
+    unpatch (hapi) {
+      this.unwrap(hapi.Server.prototype, 'start')
+      this.unwrap(hapi.Server.prototype, 'ext')
+    }
+  },
+  {
+    name: 'hapi',
+    versions: ['^7.2'],
+    patch (hapi, tracer, config) {
+      this.wrap(hapi, 'createServer', createWrapServer(tracer, config))
+    },
+    unpatch (hapi) {
+      this.unwrap(hapi, 'createServer')
+    }
+  },
+  {
+    name: '@hapi/hapi',
+    versions: ['>=17.9'],
+    file: 'lib/core.js',
+    patch (Core, tracer, config) {
+      this.wrap(Core.prototype, '_dispatch', createWrapDispatch(tracer, config))
+    },
+    unpatch (Core) {
+      this.unwrap(Core.prototype, '_dispatch')
+    }
+  },
+  {
+    name: 'hapi',
+    versions: ['7.2 - 16'],
+    file: 'lib/connection.js',
+    patch (Connection, tracer, config) {
+      this.wrap(Connection.prototype, '_dispatch', createWrapDispatch(tracer, config))
+    },
+    unpatch (Connection) {
+      this.unwrap(Connection.prototype, '_dispatch')
+    }
+  },
+  {
+    name: 'hapi',
+    versions: ['>=17'],
+    file: 'lib/core.js',
+    patch (Core, tracer, config) {
+      this.wrap(Core.prototype, '_dispatch', createWrapDispatch(tracer, config))
+    },
+    unpatch (Core) {
+      this.unwrap(Core.prototype, '_dispatch')
+    }
+  },
+  {
+    name: 'hapi',
+    versions: ['2 - 7.1'],
+    file: 'lib/server.js',
+    patch (Server, tracer, config) {
+      this.wrap(Server.prototype, '_dispatch', createWrapDispatch(tracer, config))
+    },
+    unpatch (Server) {
+      this.unwrap(Server.prototype, '_dispatch')
+    }
+  }
+]
